Restructure incremental input stack implementations to reduce code duplication
