+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/inputstack/incremental/CloneableStack.java
@@ -1,0 +1,15 @@
+package org.spoofax.jsglr2.inputstack.incremental;
+
+import java.util.Stack;
+
+class CloneableStack<E> extends Stack<E> implements IStack<E> {
+
+    @Override public CloneableStack<E> clone() {
+        CloneableStack<E> clone = new CloneableStack<>();
+        for(E e : this) {
+            clone.push(e);
+        }
+        return clone;
+    }
+
+}
