+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/inputstack/incremental/IIncrementalInputStack.java
@@ -8,7 +8,7 @@ public interface IIncrementalInputStack extends IInputStack {
 
     void breakDown();
 
-    /** Returns whether the lookahead of the node returned by {@link this#getNode} is unchanged. */
+    /** Returns whether the lookahead of the node returned by {@link #getNode()} is unchanged. */
     boolean lookaheadIsUnchanged();
 
     IIncrementalInputStack clone();
