+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/inputstack/incremental/IStack.java
@@ -1,0 +1,13 @@
+package org.spoofax.jsglr2.inputstack.incremental;
+
+interface IStack<E> {
+    E push(E e);
+
+    E pop();
+
+    E peek();
+
+    boolean isEmpty();
+
+    IStack<E> clone();
+}
