+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/inputstack/incremental/IncrementalInputStackFactory.java
@@ -4,6 +4,6 @@ import org.spoofax.jsglr2.incremental.parseforest.IncrementalParseForest;
 
 public interface IncrementalInputStackFactory<InputStack extends IIncrementalInputStack> {
 
-    InputStack get(IncrementalParseForest updatedTree, String inputString);
+    InputStack get(String inputString, String previousInput, IncrementalParseForest previousResult);
 
 }
