+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/inputstack/incremental/LazyPreprocessingIncrementalInputStack.java
@@ -1,12 +1,18 @@
 package org.spoofax.jsglr2.inputstack.incremental;
 
 import static org.spoofax.jsglr2.incremental.parseforest.IncrementalCharacterNode.EOF_NODE;
+import static org.spoofax.jsglr2.inputstack.incremental.AbstractPreprocessingIncrementalInputStack.preProcessParseForest;
 
 import java.util.Stack;
 
+import org.spoofax.jsglr2.incremental.IIncrementalParseState;
+import org.spoofax.jsglr2.incremental.diff.IStringDiff;
+import org.spoofax.jsglr2.incremental.diff.ProcessUpdates;
 import org.spoofax.jsglr2.incremental.parseforest.*;
+import org.spoofax.jsglr2.parser.AbstractParseState;
+import org.spoofax.jsglr2.stack.IStackNode;
 
-public class LazyIncrementalInputStack extends AbstractInputStack implements IIncrementalInputStack {
+public class LazyPreprocessingIncrementalInputStack extends AbstractInputStack implements IIncrementalInputStack {
     /**
      * The stack contains the parent and child index of the node that has been returned last time. When the stack is
      * initialized, a mock root is created and pushed to the stack.
@@ -18,7 +24,7 @@ public class LazyIncrementalInputStack extends AbstractInputStack implements IIn
      * @param inputString
      *            should be equal to the yield of the root.
      */
-    public LazyIncrementalInputStack(IncrementalParseForest root, String inputString) {
+    public LazyPreprocessingIncrementalInputStack(IncrementalParseForest root, String inputString) {
         super(inputString);
         IncrementalParseNode ultraRoot = new IncrementalParseNode(root, IncrementalCharacterNode.EOF_NODE);
         stack.push(new StackTuple(ultraRoot, 0));
@@ -26,12 +32,15 @@ public class LazyIncrementalInputStack extends AbstractInputStack implements IIn
         this.last = root;
     }
 
-    LazyIncrementalInputStack(IncrementalParseForest root) {
-        this(root, root.getYield());
+    public static <StackNode extends IStackNode, ParseState extends AbstractParseState<IIncrementalInputStack, StackNode> & IIncrementalParseState>
+        IncrementalInputStackFactory<IIncrementalInputStack>
+        factory(IStringDiff diff, ProcessUpdates<StackNode, ParseState> processUpdates) {
+        return (inputString, previousInput, previousResult) -> new LazyPreprocessingIncrementalInputStack(
+            preProcessParseForest(processUpdates, diff, inputString, previousInput, previousResult), inputString);
     }
 
-    @Override public LazyIncrementalInputStack clone() {
-        LazyIncrementalInputStack clone = new LazyIncrementalInputStack(EOF_NODE, inputString);
+    @Override public LazyPreprocessingIncrementalInputStack clone() {
+        LazyPreprocessingIncrementalInputStack clone = new LazyPreprocessingIncrementalInputStack(EOF_NODE, inputString);
         clone.stack.clear();
         for(StackTuple stackTuple : stack) {
             clone.stack.push(stackTuple);
@@ -83,9 +92,9 @@ public class LazyIncrementalInputStack extends AbstractInputStack implements IIn
     }
 
     @Override public boolean lookaheadIsUnchanged() {
-        for (int i = stack.size() - 1; i >= 0; i--) {
+        for(int i = stack.size() - 1; i >= 0; i--) {
             IncrementalParseForest node = rightSibling(stack.get(i));
-            if(node != null){
+            if(node != null) {
                 if(node.isTerminal())
                     return true;
                 return ((IncrementalParseNode) node).production() != null;
