+++ org.spoofax.jsglr2/src/test/java/org/spoofax/jsglr2/inputstack/incremental/EagerPreprocessingIncrementalInputStackTest.java
@@ -2,10 +2,10 @@ package org.spoofax.jsglr2.inputstack.incremental;
 
 import org.spoofax.jsglr2.incremental.parseforest.IncrementalParseNode;
 
-public class EagerIncrementalInputStackTest extends AbstractIncrementalInputStackTest {
+public class EagerPreprocessingIncrementalInputStackTest extends AbstractPreprocessingIncrementalInputStackTest {
 
     @Override protected IIncrementalInputStack getStack(IncrementalParseNode root, String inputString) {
-        return new EagerIncrementalInputStack(root, inputString);
+        return new EagerPreprocessingIncrementalInputStack(root, inputString);
     }
 
 }
