+++ org.spoofax.jsglr2/src/test/java/org/spoofax/jsglr2/inputstack/incremental/IncrementalInputStackTest.java
@@ -5,11 +5,11 @@ import java.util.Arrays;
 import org.spoofax.jsglr2.incremental.EditorUpdate;
 import org.spoofax.jsglr2.incremental.parseforest.IncrementalParseNode;
 
-public class LinkedIncrementalInputStackTest extends AbstractIncrementalInputStackTest {
+public class IncrementalInputStackTest extends AbstractIncrementalInputStackTest {
 
     @Override protected IIncrementalInputStack getStack(IncrementalParseNode previousResult, String inputString,
         EditorUpdate... editorUpdates) {
-        return new LinkedIncrementalInputStack(inputString, previousResult, Arrays.asList(editorUpdates));
+        return new IncrementalInputStack(inputString, previousResult, Arrays.asList(editorUpdates));
     }
 
 }
