+++ org.spoofax.jsglr2/src/test/java/org/spoofax/jsglr2/inputstack/incremental/LazyPreprocessingIncrementalInputStackTest.java
@@ -1,0 +1,11 @@
+package org.spoofax.jsglr2.inputstack.incremental;
+
+import org.spoofax.jsglr2.incremental.parseforest.IncrementalParseNode;
+
+public class LazyPreprocessingIncrementalInputStackTest extends AbstractPreprocessingIncrementalInputStackTest {
+
+    @Override protected IIncrementalInputStack getStack(IncrementalParseNode root, String inputString) {
+        return new LazyPreprocessingIncrementalInputStack(root, inputString);
+    }
+
+}
