+++ org.spoofax.jsglr2/src/test/java/org/spoofax/jsglr2/inputstack/incremental/LinkedPreprocessingIncrementalInputStackTest.java
@@ -2,10 +2,10 @@ package org.spoofax.jsglr2.inputstack.incremental;
 
 import org.spoofax.jsglr2.incremental.parseforest.IncrementalParseNode;
 
-public class LazyIncrementalInputStackTest extends AbstractIncrementalInputStackTest {
+public class LinkedPreprocessingIncrementalInputStackTest extends AbstractPreprocessingIncrementalInputStackTest {
 
     @Override protected IIncrementalInputStack getStack(IncrementalParseNode root, String inputString) {
-        return new LazyIncrementalInputStack(root, inputString);
+        return new LinkedPreprocessingIncrementalInputStack(root, inputString);
     }
 
 }
